Split UserInfoProvider into separate files, renamed UserInfo to Driver
